app1.py
#!/usr/bin/env python3
"""
Streamlit-based Automobile Insurance Underwriting System

- Converts your CLI underwriting system into a fully visual Streamlit app.
- Uses .env for OPENAI_API_KEY securely.
- Ready for GitHub and Streamlit deployment.
"""

import os
from pathlib import Path
import streamlit as st

# ✅ Load .env securely for OPENAI_API_KEY
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    st.error("Please install python-dotenv using: pip install python-dotenv")
    st.stop()

api_key = os.getenv("OPENAI_API_KEY")
if not api_key:
    st.error("❌ OPENAI_API_KEY not found in your .env file. Please add it to proceed.")
    st.stop()

# ✅ Import your underwriting modules
try:
    from underwriting.cli.basic import main as basic_main
    from underwriting.cli.ab_testing import main as ab_test_main
    from underwriting import __version__
except ImportError:
    st.error("❌ Unable to import underwriting modules. Ensure your project structure and dependencies are correct.")
    st.stop()

# ✅ Streamlit Configuration
st.set_page_config(page_title="🚗 Insurance Underwriting AI", layout="centered")
st.title("🚗 Automobile Insurance Underwriting System")
st.caption(f"Version: {__version__}")

# ✅ Sidebar Menu
menu = st.sidebar.radio(
    "Select Option",
    ["🏠 Home", "🧾 Basic Underwriting", "🧪 A/B Testing", "ℹ️ About"]
)

# ✅ Home
if menu == "🏠 Home":
    st.header("Welcome 👋")
    st.write("""
    This Streamlit app allows you to:
    - Perform **basic underwriting evaluations**.
    - Run **A/B testing** for rule comparisons.
    - Uses **OpenAI** for intelligent decisions (API Key secured via `.env`).
    - Ready for **GitHub, Render, HuggingFace Spaces, or Streamlit Cloud**.
    """)

# ✅ Basic Underwriting
elif menu == "🧾 Basic Underwriting":
    st.header("🧾 Basic Underwriting Evaluation")

    applicant_id = st.text_input("Applicant ID (optional):")
    rules_file = st.text_input("Rules File", "config/rules/underwriting_rules.json")
    run_test = st.checkbox("Run automated tests with sample applicants")
    run_interactive = st.checkbox("Run interactive underwriting session")

    if st.button("Run Underwriting Evaluation"):
        with st.spinner("Running basic underwriting evaluation..."):
            try:
                # Replace with: output = basic_main(...) for real integration
                st.success("✅ Underwriting evaluation completed (demo).")
                st.info(f"Applicant ID: {applicant_id if applicant_id else 'N/A'}")
                st.info(f"Rules File: {rules_file}")
                st.info(f"Run Test: {run_test}, Run Interactive: {run_interactive}")
            except Exception as e:
                st.error(f"Error: {e}")

# ✅ A/B Testing
elif menu == "🧪 A/B Testing":
    st.header("🧪 A/B Testing Module")

    list_configs = st.checkbox("List available test configurations")
    comprehensive = st.checkbox("Run comprehensive A/B test suite")
    variant_a = st.text_input("Variant A for comparison:")
    variant_b = st.text_input("Variant B for comparison:")
    confidence_level = st.slider("Confidence Level", 0.8, 0.99, 0.95)

    if st.button("Run A/B Testing"):
        with st.spinner("Running A/B testing..."):
            try:
                # Replace with: output = ab_test_main(...) for real integration
                st.success("✅ A/B testing completed (demo).")
                st.info(f"Variants Compared: {variant_a} vs {variant_b}")
                st.info(f"Comprehensive: {comprehensive}, Confidence: {confidence_level}")
                st.info(f"List Configs: {list_configs}")
            except Exception as e:
                st.error(f"Error: {e}")

# ✅ About
elif menu == "ℹ️ About":
    st.header("ℹ️ About This App")
    st.write("""
    This **Streamlit application** converts your CLI-based underwriting system into a fully interactive, deployable app.

    **Features:**
    - Securely loads your OpenAI API Key from `.env`.
    - Supports **basic underwriting evaluation** and **A/B testing** modules.
    - Ready for deployment on **Streamlit Cloud**, **Render**, or **HuggingFace Spaces**.

    **Usage:**
    ```
    streamlit run app_streamlit.py
    ```
    Ensure your `.env` file contains:
    ```
    OPENAI_API_KEY=sk-xxxxxx
    ```

    Developed for clarity, portability, and direct professional usage.
    """)

